+++ Test_Spokenizer.py
@@ -1,0 +1,19 @@
+import unittest
+from Posterior import *
+
+class Test_tokenizer(unittest.TestCase):
+
+    def test_what_it_receives_is_the_initial_sentence_property(self):
+        pass
+    def test_that_domainer_was_created(self):
+        pass
+    def test_that_emotionizer_was_created(self):
+        pass
+    def test_what_it_outputs_is_a_list(self):
+        pass
+    # we also need tests in here that output the word count, character count, word placement
+
+
+if __name__ == '__main__':
+    unittest.main()
+
